refactor: move export tests to separate export_test.rs module
Relocate the inline test module from export.rs into a sibling
export_test.rs, matching the convention used in src/common/config/
(template_test.rs, variables_test.rs). Keeps tests out of the
implementation file.
